bugfix in traffic limit and whitelist

+++ src/de/szalkowski/adamsbatterysaver/MobileDataPowerSaver.java
@@ -58,7 +58,7 @@ public class MobileDataPowerSaver extends PowerSaver {
 			this.traffic = TrafficStats.getMobileRxBytes() + TrafficStats.getMobileTxBytes();
 			
 	        SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(context.getApplicationContext());
-			final long traffic_limit = settings.getLong(MainActivity.SETTINGS_TRAFFIC_LIMIT, (long)context.getResources().getInteger(R.integer.pref_traffic_limit_default));
+			final long traffic_limit = (long)settings.getInt(MainActivity.SETTINGS_TRAFFIC_LIMIT, context.getResources().getInteger(R.integer.pref_traffic_limit_default));
 			final double traffic_per_minute = traffic_diff/(time_diff/60000.0);
 			Log.v(LOG,"mobile traffic: " + traffic_per_minute + " bytes / minute ("+ traffic_diff + "/" + time_diff/1000.0 + "s)");
 			if(traffic_per_minute > traffic_limit) {
