+++ src/de/szalkowski/adamsbatterysaver/WifiPowerSaver.java
@@ -59,7 +59,7 @@ public class WifiPowerSaver extends PowerSaver {
 			this.traffic = TrafficStats.getTotalRxBytes() + TrafficStats.getTotalTxBytes() - TrafficStats.getMobileRxBytes() - TrafficStats.getMobileTxBytes();
 			
 	        SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(context.getApplicationContext());
-			final long traffic_limit = settings.getLong(MainActivity.SETTINGS_TRAFFIC_LIMIT, (long)context.getResources().getInteger(R.integer.pref_traffic_limit_default));
+			final long traffic_limit = (long)settings.getInt(MainActivity.SETTINGS_TRAFFIC_LIMIT, context.getResources().getInteger(R.integer.pref_traffic_limit_default));
 			final double traffic_per_minute = traffic_diff/(time_diff/60000.0);
 			Log.v(LOG,"wifi traffic: " + traffic_per_minute + " bytes / minute ("+ traffic_diff + "/" + time_diff/1000.0 + "s)");
 			if(traffic_per_minute > traffic_limit) {
@@ -70,6 +70,7 @@ public class WifiPowerSaver extends PowerSaver {
 			List<ActivityManager.RunningTaskInfo> tasks = am.getRunningTasks(1);
 			if(!tasks.isEmpty()) {
 				String currentTaskPackage = tasks.get(0).topActivity.getPackageName();
+				
 				Log.d(LOG,"current task: " + currentTaskPackage);
 				if(settings.getStringSet("wifi_whitelist", new HashSet<String>()).contains(currentTaskPackage)) {
 					Log.d(LOG,currentTaskPackage + " is on whitelist");
